+++ java/com/jcraft/jsch/Session.java
@@ -1,7 +1,7 @@
 /* -*-mode:java; c-basic-offset:2; indent-tabs-mode:nil -*- */
 /*
 Copyright (c) 2002-2018 ymnk, JCraft,Inc. All rights reserved.
-Copyright (c) 2018, 2023-2024 D. R. Commander. All rights reserved.
+Copyright (c) 2018, 2023-2025 D. R. Commander. All rights reserved.
 Copyright (c) 2020-2021 Jeremy Norris. All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
@@ -168,6 +168,8 @@ public class Session implements Runnable{
     this.port = port;
 
     applyConfig();
+
+    this.host = this.host.replaceAll("[\\[\\]]", "");
 
     if(this.username==null) {
       try {